requirements.txt
gradio
langchain
langchain_community
langchain_openai
beautifulsoup4
kss
chromadb
langchain-upstage
python-dotenv
requests
fastapi
uvicorn
langchain-text-splitters
tiktoken
sentence-transformers
